LayoutTests/fast/dom/CSSStyleDeclaration/access-longest-css-property.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<style>
* {
-webkit-border-bottom-right-radius: 5px;
}
</style>
</head>
<body>
<div id="target" style="-webkit-border-bottom-right-radius: 5px;;"></div>
<!--
Note for updating the test:
Please replace '-webkit-border-bottom-right-radius' by the longest name if '-webkit-border-bottom-right-radius' is removed.
-->
<script>
description("The test parsing of the boundary value by checking the longest CSS property.");
shouldBeEqualToString('getComputedStyle(document.body).webkitBorderBottomRightRadius', '5px');
shouldBeEqualToString('document.getElementById("target").style.webkitBorderBottomRightRadius', '5px');
shouldBeEqualToString('document.body.style.webkitBorderBottomRightRadius', '');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
